refactor: extract focusable directive into module
