+++ apps/rockket-web/src/app/app.module.ts
@@ -47,10 +47,10 @@ import { SidepanelPortalDirective } from './components/templates/main-pane/sidep
 import { MenuToggleComponent } from './components/templates/sidebar-layout/menu-toggle/menu-toggle.component'
 import { SidebarLayoutComponent } from './components/templates/sidebar-layout/sidebar-layout.component'
 import { DiffModule } from './diff/diff.module'
-import { FocusableDirective } from './directives/focusable.directive'
 import { IntersectionDirective } from './directives/intersection.directive'
 import { MutationDirective } from './directives/mutation.directive'
 import { DropdownModule } from './dropdown/dropdown.module'
+import { FocusableModule } from './focusable/focusable.module'
 import { HttpModule } from './http/http.module'
 import { KeyboardModule } from './keyboard/keyboard.module'
 import { ModalModule } from './modal/modal.module'
@@ -91,7 +91,6 @@ import { TooltipModule } from './tooltip/tooltip.module'
         DemoComponent,
         TaskComponent,
         ComponentPlaygroundComponent,
-        FocusableDirective,
         InputComponent,
         FormComponent,
         LoginComponent,
@@ -171,6 +170,7 @@ import { TooltipModule } from './tooltip/tooltip.module'
         PortalModule,
         DiffModule,
         ResizeModule,
+        FocusableModule,
     ],
     providers: [
         {
