+++ apps/rockket-web/src/app/components/molecules/editable-entity-heading/editable-entity-title.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing'
-import { FocusableDirective } from 'src/app/directives/focusable.directive'
+import { FocusableDirective } from 'src/app/focusable/focusable.directive'
 import { actionsMock, storeMock } from 'src/app/utils/unit-test.mocks'
 
 import { EditableEntityTitleComponent } from './editable-entity-title.component'
