+++ apps/rockket-web/src/app/components/organisms/task/task.component.cy.ts
@@ -2,7 +2,7 @@ import { CdkMenuModule } from '@angular/cdk/menu'
 import { Store } from '@ngrx/store'
 import { EntityType, TaskFlattend, TaskPriority, TaskStatus } from '@rockket/commons'
 import { testName, useStubsForActions } from 'cypress/support/helpers'
-import { FocusableDirective } from 'src/app/directives/focusable.directive'
+import { FocusableDirective } from 'src/app/focusable/focusable.directive'
 import { MutationDirective } from 'src/app/directives/mutation.directive'
 import { DropdownModule } from 'src/app/dropdown/dropdown.module'
 import { HighlightPipe } from 'src/app/pipes/highlight.pipe'
