+++ apps/rockket-web/src/app/components/organisms/task/task.component.spec.ts
@@ -2,7 +2,7 @@ import { CdkMenuModule } from '@angular/cdk/menu'
 import { ComponentFixture, TestBed } from '@angular/core/testing'
 import { TaskFlattend, TaskPriority, TaskStatus } from '@rockket/commons'
 import { BehaviorSubject } from 'rxjs'
-import { FocusableDirective } from 'src/app/directives/focusable.directive'
+import { FocusableDirective } from 'src/app/focusable/focusable.directive'
 import { HighlightPipe } from 'src/app/pipes/highlight.pipe'
 import { RxModule } from 'src/app/rx/rx.module'
 import { IconsModule } from '../../atoms/icons/icons.module'
